fix(electron-app): Make the binary work correctly (#316)
* Add menu, use MemoryRouter

* Default to alex

* Add header

* Don't default to alex

* Check before kill (fixes #317)

+++ packages/electron-app/src/main/util/substrateProcess.ts
@@ -31,8 +31,10 @@ export const runSubstrateDev = () => {
 };
 
 export const killSubstrate = () => {
-  pino.info('setting substrate process -> ', substrateProc);
-  substrateProc.kill();
+  if (substrateProc) {
+    pino.info('setting substrate process -> ', substrateProc);
+    substrateProc.kill();
+  }
 };
 
 export const purgeDevChain = () => {
